hw1/requirements.txt
pandas
sqlalchemy
psycopg2-binary
matplotlib
seaborn

